ai-gateway/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
